refact: Validator base element
