+++ panel/vite.config.ts
@@ -71,7 +71,8 @@ function createPlugins(mode: string): Plugin[] {
 		vue({
 			template: {
 				compilerOptions: {
-					isCustomElement: (tag) => ["k-input-validator"].includes(tag)
+					isCustomElement: (tag) =>
+						["k-input-validator", "k-validator"].includes(tag)
 				}
 			}
 		}),
